pyproject.toml
[tool.poetry]
name = "openai-lambda"
version = "0.1.0"
description = "Lambda Functions with OpenAI"
authors = ["Xanthous Tech <hi@x-tech.io>"]
license = "ISC"

[tool.poetry.dependencies]
python = "^3.8"
openai = "^0.6.4"
python-dotenv = "^0.17.1"
boto3 = "^1.17.89"
jwcrypto = "^0.9"
aws-lambda-powertools = "^1.16.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
